feat: Modulos separados para las diferentes funciones
