+++ modules/base64_to_numpy.py
@@ -1,0 +1,20 @@
+from PIL import Image
+from io import BytesIO
+
+import base64
+import face_recognition
+import numpy as np
+
+# Transformamos de base64 a numpy array
+def base64_to_numpy(image):
+    image_bytes = base64.b64decode(image)
+    # Crear un objeto BytesIO a partir de los bytes
+    image_io = BytesIO(image_bytes)
+    # Abrir la imagen utilizando la biblioteca PIL (Pillow)
+    image = Image.open(image_io)
+    # Convertir la imagen a una matriz numpy y encodear
+    img_array = np.array(image)
+    return {
+        "image_encode": face_recognition.face_encodings(img_array)[0],
+    }
+
